Tried to fix bug with cache expiry from project observations.

+++ app/models/project_observation.rb
@@ -246,9 +246,10 @@ class ProjectObservation < ActiveRecord::Base
   end
 
   def expire_caches
+    return true if project_id.blank?
     begin
       FileUtils.rm private_page_cache_path(FakeView.all_project_observations_path(project, :format => 'csv')), :force => true
-    rescue ActionController::RoutingError
+    rescue ActionController::RoutingError, ActionController::UrlGenerationError
       FileUtils.rm private_page_cache_path(FakeView.all_project_observations_path(project_id, :format => 'csv')), :force => true
     end
     true
